Add important changes

+++ pylan/shell/rs_conf.sh
@@ -5,14 +5,15 @@ wpconf="/etc/wpa_supplicant/wpa_supplicant-$WLAN.conf"
 wlan_network="/etc/systemd/network/*-$WLAN.network"
 
 rm -rf $wpconf $wlan_network
-systemctl unmask networking.service dhcpcd.service
-mv /etc/network/{interfaces~,interfaces} # restore file
-systemctl disable systemd-networkd.service systemd-resolved.service wpa_supplicant@$WLAN.service
-systemctl enable wpa_supplicant.service
+sudo systemctl unmask networking.service dhcpcd.service
+if [[ -f "/etc/network/interfaces~" ]]; then
+    mv /etc/network/{interfaces~,interfaces}
+fi
+sudo systemctl disable systemd-networkd.service systemd-resolved.service wpa_supplicant@$WLAN.service
+sudo systemctl enable wpa_supplicant.service
 rm -rf /etc/resolv.conf
 cp /etc/{resolv.conf~,resolv.conf}
 
-sudo systemctl restart systemd-networkd.service
 sudo systemctl restart wpa_supplicant.service
 
 echo "1"
